requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.2.0
tqdm>=4.61.0
scikit-learn>=0.24.0
faiss-cpu>=1.7.2

# Transformers for DiNO and CLIP models
transformers>=4.21.0

# Visualization tools
matplotlib>=3.4.0

